Create crisp
Crisp is a crypto- and finance-driven site. The site aggregates content from YouTube, Spotify and other content creation media platforms. The purpose of the platform is to drive user interaction and encourage content creators to create material exclusively for the site. This is a NodeJS platform built with VueJS. It integrates several APIs including the YouTube data API, TradingView and Google signin. The site employs a MySQL database that tracks client usage, records tendencies, recommends new content and builds a profile of the user.